requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
PyMySQL==1.1.2
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
zipp==3.23.0
